Undo renaming.
