+++ lib/open-uri-s3.rb
