+++ lib/open-uri-s3/version.rb
