+++ open-uri-s3.gemspec
@@ -3,7 +3,7 @@
 lib = File.expand_path('../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
 
-require 'open_uri_s3/version'
+require 'open-uri-s3/version'
 
 Gem::Specification.new do |s|
   s.name        = 'open-uri-s3'
